#5 - Develop permissions concept Move my-authorities to auth file
