+++ server/routes/auth.ts
@@ -6,6 +6,9 @@ import passport, {AuthenticateOptions} from 'passport';
 import {ErrorBody} from "../typings/error";
 import jwtMiddleware from '../middlewares/jwt-middleware'
 import {IUser} from "../typings/user";
+import {getAllRoles} from "../models/role";
+import {getAuthoritiesByUserId} from "../models/authorities";
+import {IUserAuthorities} from "../typings/authorities";
 
 const router = Router();
 
@@ -157,6 +160,28 @@ router.get('/me', jwtMiddleware, (req: Request, res: Response) => {
         });
 });
 
+
+router.get('/my-authorities', jwtMiddleware, (req: Request, res: Response) => {
+    const userId = (req.user as any)._id;
+    Promise.all([
+        getAllRoles(),
+        getAuthoritiesByUserId(userId)
+    ])
+        .then(([roles, authorities]) => {
+            const userAuths: IUserAuthorities[] = authorities.map(a => {
+                const role = roles.find(r => r.name === a.role);
+                return { ...a, permissions: role?.permissions }
+            });
+            res.send(userAuths);
+        })
+        .catch(err => {
+            const errorBody: ErrorBody = {
+                message: 'Failed to get user permissions',
+                details: err
+            };
+            res.status(500).json(errorBody);
+        })
+});
 
 /**
  * Get list of available external authentication services (e.g. Google, LinkedIn, Facebook)
